+++ app/auth/user_page/project/page.tsx
@@ -15,7 +15,7 @@ async function getData() {
     );
     if (response.ok) {
       const data = await response.json();
-      console.log(data);
+      // console.log(data);
       return data;
     }
   } catch (error) {
@@ -29,9 +29,6 @@ export default async function ProjectComponentShow() {
     // Xử lý trường hợp không có dự án nào được trả về từ API
     return <div>No projects found.</div>;
   }
-  // const projectData = data[0];
-  // return <ProjectComponent data={projectData} />;
-
   return (
     <div className="grid grid-cols-2 grid-rows-2 sm:grid-cols-2 md:grid-cols-2 gap-4">
       {data.map(
@@ -40,8 +37,8 @@ export default async function ProjectComponentShow() {
             projectID: number;
             projectName: string;
             projectSlug: string;
-            totalLink: string;
-            totalClick: string;
+            totalLink: number;
+            totalClick: number;
           },
           index: number,
         ) => (
