+++ components/auth/projectpage/create-link.tsx
@@ -1,3 +1,4 @@
+"use client";
 import { Button } from "@/components/ui/button";
 import {
   Dialog,
